pyproject.toml
[tool.poetry]
name = "webhook"
version = "0.1.0"
description = ""
authors = ["Jose"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
pandas = "^2.1.2"
amqp = "^5.2.0"
beautifulsoup4 = "^4.12.2"
celery = "^5.3.4"
colorama = "^0.4.6"
django = "^4.2.7"
djangorestframework = "^3.14.0"
drf-yasg = "^1.21.7"
flower = "^2.0.1"
gunicorn = "^21.2.0"
heroku3 = "^5.2.1"
httpx = "^0.25.1"
inflection = "^0.5.1"
openpyxl = "^3.1.2"
pdfminer = "^20191125"
pika = "^1.3.2"
python-dotenv = "^1.0.0"
psycopg = "^3.1.12"
psycopg2-binary = "^2.9.9"
whitenoise = "^6.6.0"
gevent = "^23.9.1"
rocketry = "^2.5.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
